Separation of logic into utils files
